+++ bioimageio/spec/_internal/common_nodes.py
@@ -222,12 +222,12 @@ class ResourceDescrBase(
             file_descrs = extract_file_descrs({k: v for k, v in data.items()})
             populate_cache(file_descrs)  # TODO: add progress bar
 
-        with context:
+        with context.replace(log_warnings=context.warning_level <= INFO):
             rd, errors, val_warnings = cls._load_impl(deepcopy_yaml_value(data))
 
         if context.warning_level > INFO:
             all_warnings_context = context.replace(
-                warning_level=INFO, log_warnings=False
+                warning_level=INFO, log_warnings=False, raise_errors=False
             )
             # raise all validation warnings by reloading
             with all_warnings_context: