+++ bioimageio/spec/application/v0_3.py
@@ -5,7 +5,7 @@ from typing_extensions import Annotated
 
 from .._internal.io import FileDescr as FileDescr
 from .._internal.io_basics import Sha256 as Sha256
-from .._internal.types import FileSource_
+from .._internal.types import FAIR, FileSource_
 from .._internal.url import HttpUrl as HttpUrl
 from ..generic.v0_3 import VALID_COVER_IMAGE_EXTENSIONS as VALID_COVER_IMAGE_EXTENSIONS
 from ..generic.v0_3 import Author as Author
@@ -46,7 +46,7 @@ class ApplicationDescr(GenericDescrBase):
     """The description from which this one is derived"""
 
     source: Annotated[
-        Optional[FileSource_],
+        FAIR[Optional[FileSource_]],
         Field(description="URL or path to the source of the application"),
     ] = None
     """The primary source of the application"""