+++ bioimageio/spec/dataset/v0_3.py
@@ -5,6 +5,7 @@ from pydantic import model_validator
 from .._internal.common_nodes import InvalidDescr
 from .._internal.io import FileDescr as FileDescr
 from .._internal.io_basics import Sha256 as Sha256
+from .._internal.types import FAIR
 from .._internal.url import HttpUrl as HttpUrl
 from ..generic.v0_3 import VALID_COVER_IMAGE_EXTENSIONS as VALID_COVER_IMAGE_EXTENSIONS
 from ..generic.v0_3 import Author as Author
@@ -53,7 +54,7 @@ class DatasetDescr(GenericDescrBase):
     parent: Optional[DatasetId] = None
     """The description from which this one is derived"""
 
-    source: Optional[HttpUrl] = None
+    source: FAIR[Optional[HttpUrl]] = None
     """"URL to the source of the dataset."""
 
     @model_validator(mode="before")